server/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
psycopg2-binary
requests
beautifulsoup4
praw
python-dotenv
gunicorn
